backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# Database
sqlalchemy==2.0.23
alembic==1.13.0
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# OpenAI integration
openai==1.3.5

# Environment and configuration
python-dotenv==1.0.0
pydantic-settings==2.1.0

# Email validation
email-validator==2.2.0

# HTTP client
httpx==0.25.2

# Monitoring (leveraging existing)
prometheus-client==0.19.0
psutil==5.9.6

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
aiosqlite==0.21.0

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0